third_party/blink/web_tests/http/tests/security/referrer-policy-conflicting-policies-image-set-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/resources/referrer-policy-conflicting-policies-image-set.html' was loaded over HTTPS, but requested an insecure element 'http://127.0.0.1:8080/security/resources/green-if-no-referrer.php'. This request was not upgraded to HTTPS because its URL's host is an IP address.
